Move InputOutputStream and SerializationUtils
