+++ platform-sdk/base-utility/build.gradle.kts
@@ -9,7 +9,6 @@ description = "Base Utility"
 
 testModuleInfo {
     requires("com.swirlds.base.test.fixtures")
-    requires("com.swirlds.common.test.fixtures")
     requires("org.hiero.base.crypto")
     requires("org.hiero.base.utility.test.fixtures")
     requires("org.assertj.core")
