+++ platform-sdk/consensus-event-stream/build.gradle.kts
@@ -10,7 +10,6 @@ description = "Consensus event-stream file writing"
 
 testModuleInfo {
     requires("com.swirlds.common")
-    requires("com.swirlds.common.test.fixtures")
     requires("com.swirlds.platform.core.test.fixtures")
     requires("org.hiero.base.crypto.test.fixtures")
     requires("org.hiero.base.utility.test.fixtures")
