+++ platform-sdk/consensus-event-stream/src/test/java/org/hiero/consensus/event/stream/StreamUtilitiesTest.java
@@ -38,7 +38,6 @@ import static org.mockito.Mockito.mock;
 
 import com.swirlds.base.utility.Pair;
 import com.swirlds.common.constructable.ConstructableRegistration;
-import com.swirlds.common.test.fixtures.io.InputOutputStream;
 import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.IOException;
@@ -61,6 +60,7 @@ import org.hiero.base.crypto.RunningHashable;
 import org.hiero.base.crypto.Signature;
 import org.hiero.base.crypto.test.fixtures.CryptoRandomUtils;
 import org.hiero.base.io.SelfSerializable;
+import org.hiero.base.utility.test.fixtures.io.InputOutputStream;
 import org.hiero.consensus.event.stream.internal.LinkedObjectStreamValidateUtils;
 import org.hiero.consensus.event.stream.internal.StreamValidationResult;
 import org.hiero.consensus.event.stream.test.fixtures.ObjectForTestStream;
